erp5_catalog: [WORKAROUND] Explicitly call reset for component_tool in test.
After adding document component object, _p_oid was bing generated at the
last step of commit hook. This led to problem that _registry_dict for
dynamic_class 'erp5.component.document' wasn't getting updated with
the _p_oid of the component created. This led to failing of test because
to validate object with same reference, the checkConsistency function
checks in the _registry_dict for reference and _p_oid.

Explicilty calling commit here generates the oid before validation, hence
make it available for the registry_dict in time.

But, this clearly is a workaround which is clearly not solving the real
problem of why the _p_oid isn't being generated at the right step.

Also, me(Ayush), tried to recreate this error on a running instance, but
it worked fine there.